% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dis_data}
\alias{dis_data}
\title{A dissolution data set taken from Ocana et al. (2009).}
\format{
A data frame with 24 rows and 9 columns:
\describe{
\item{group}{An indicator of whether the dissolution run came from the reference or test group}
\item{X1}{The first time point at which measurements are made at.}
\item{X2}{The second time point at which measurements are made at.}
\item{X3}{The third time point at which measurements are made at.}
\item{X4}{The fourth time point at which measurements are made at.}
\item{X5}{The fifth time point at which measurements are made at.}
\item{X6}{The sixth time point at which measurements are made at.}
\item{X7}{The seventh time point at which measurements are made at.}
\item{X8}{The eight time point at which measurements are made at.}
...
}
}
\source{
Ocana et al. (2009) \url{doi:10.1016/j.chemolab.2009.07.010}
}
\usage{
dis_data
}
\description{
A dissolution data set that consists of dissolution measurements
taken on oral tablets made with metoclopramide hydrochloride. Of interest is
to test the similarity of metoclopramide hydrochloride tablets made with and without the
ingredient tensioactive.
}
\keyword{datasets}
