% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dis_data.R
\name{make_dis_data}
\alias{make_dis_data}
\title{Class dis_data creation}
\usage{
make_dis_data(yRef, yTest)
}
\arguments{
\item{yRef}{A data frame or matrix containing the dissolution data for the reference group data. The rows of the data set correspond
to the individual dissolution runs. The columns of the data frame contains the individual run's dissolution
measurements sorted in time.}

\item{yTest}{A data frame or matrix containing the dissolution data for the test group data. The rows of the data set correspond
to the individual dissolution runs. The columns of the data frame contains the individual run's dissolution
measurements sorted in time.}
}
\value{
The function returns a data object of class dis_data.
}
\description{
This function creates a data object of class dis_data.
}
\examples{
### dis_data comes loaded with the package
### but need to update dis_data to be an object of class dis_data
new_dis_data <- make_dis_data(yRef = dis_data[dis_data$group == "Reference", -1],
                              yTest = dis_data[dis_data$group == "Test", -1])

}
